Fix vercel.json for both backend and frontend and cors issues
