+++ backend/.env.sample
@@ -1,5 +1,7 @@
+PORT=8080
 MONGODB_URI="mongodb://127.0.0.1/wanderlust"
 REDIS_URL="redis://127.0.0.1:6379"
+FRONTEND_URL=http://localhost:5173
 ACCESS_COOKIE_MAXAGE=120000
 ACCESS_TOKEN_EXPIRES_IN='120s'
 REFRESH_COOKIE_MAXAGE=120000
