+++ backend/api/index.js
@@ -1,2 +1,2 @@
-import app from '../server.js';
+import app from '../app.js';
 export default app;
