+++ backend/app.js
@@ -1,0 +1,42 @@
+import compression from 'compression';
+import cookieParser from 'cookie-parser';
+import cors from 'cors';
+import express from 'express';
+import { FRONTEND_URL } from './config/utils.js';
+import authRouter from './routes/auth.js';
+import postsRouter from './routes/posts.js';
+import userRouter from './routes/user.js';
+
+
+const app = express();
+
+app.use(cors({
+    // added origin
+    origin: FRONTEND_URL,
+    credentials: true
+}));
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.use(cookieParser());
+app.use(compression());
+
+
+// API route
+app.use('/api/posts', postsRouter);
+app.use('/api/auth', authRouter);
+app.use('/api/user', userRouter);
+
+app.get('/', (req, res) => {
+    res.send('Yay!! Backend of wanderlust app is now accessible');
+});
+
+app.all("*", (req, res) => {
+    res.status(404).json({
+        status: 404,
+        success: false,
+        message: "!Oops page not found"
+    })
+})
+
+
+export default app;
