+++ backend/config/db.js
@@ -1,8 +1,13 @@
 import mongoose from 'mongoose';
 import { MONGODB_URI } from './utils.js';
-export default function connectDB() {
+
+export default async function connectDB() {
   try {
-    mongoose.connect(MONGODB_URI);
+    await mongoose.connect(MONGODB_URI, {
+      // set database name
+      dbName: 'wanderlust'
+    });
+    console.log(`Database connected: ${MONGODB_URI}`);
   } catch (err) {
     console.error(err.message);
     process.exit(1);
@@ -10,12 +15,8 @@ export default function connectDB() {
 
   const dbConnection = mongoose.connection;
 
-  dbConnection.once('open', () => {
-    console.log(`Database connected: ${MONGODB_URI}`);
-  });
-
   dbConnection.on('error', (err) => {
-    console.error(`connection error: ${MONGODB_URI}`);
+    console.error(`connection error: ${err}`);
   });
   return;
 }
