+++ backend/server.js
@@ -1,39 +1,20 @@
-import compression from 'compression';
-import cookieParser from 'cookie-parser';
-import cors from 'cors';
-import express from 'express';
-import connectDB from './config/db.js';
-import { FRONTEND_URL, PORT } from './config/utils.js';
-import authRouter from './routes/auth.js';
-import postsRouter from './routes/posts.js';
-import userRouter from './routes/user.js';
-import { connectToRedis } from './services/redis.js';
-const app = express();
-const port = PORT || 5000;
+import app from "./app.js";
+import connectDB from './config/db.js'
+import { PORT } from "./config/utils.js";
+import { connectToRedis } from "./services/redis.js";
 
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use(cors());
-app.use(cookieParser());
-app.use(compression());
-
-// Connect to database
-connectDB();
+const port = PORT || 5000
 
 // Connect to redis
-connectToRedis();
+connectToRedis()
 
-// API route
-app.use('/api/posts', postsRouter);
-app.use('/api/auth', authRouter);
-app.use('/api/user', userRouter);
-
-app.get('/', (req, res) => {
-  res.send('Yay!! Backend of wanderlust app is now accessible');
-});
-
-app.listen(port, () => {
-  console.log(`Server is running on port ${port}`);
-});
-
-export default app;
+//Connect to Mongodb
+connectDB()
+  .then(() => {
+    app.listen(port, () => {
+      console.log(`Server is running on port ${port}`);
+    });
+  })
+  .catch((error) => {
+    console.log("MongoDB connection failed:", error);
+  })
